Feature/#11 create post

+++ internal/global/constant.go
@@ -1,5 +1,5 @@
 package global
 
 var (
-	SESSION_EXPIRE_IN_SECOND int = 3600
+	SESSION_EXPIRE_IN_SECOND int64 = 3600
 )
